chore: cg old version
